Voting Template Changed with tags for talks
Now the voting.html template (used in `conference.views.voting` callback) includes only the set of tags associated to talks
of the current conference (thanks to the new `p3_tags_for_talks` template tag).

The template has been changed trying to not break backward compatibilities with other templates extending "base.html".
To this end, the "base.html" template now wraps the invocation to `p3_tags` and `conference_js_data` tags in a new block that
is overridden by the `voting.html` template.

Finally, the CSS style for tags has been modified as well in order to allow for a better usability of tags filter.